+++ packages/client/locales/en/calls.json
@@ -33,6 +33,7 @@
     "callTow": "Call Tow",
     "departments": "Departments",
     "divisions": "Divisions",
+    "callFilters": "Call Filters",
     "alert_deleteCallEvent": "Are you sure you want to delete this event? This action cannot be undone.",
     "alert_end911Call": "Are you sure you want to end this call?",
     "alert_endTowCall": "Are you sure you want to end this call?"
